changed login to static path
