+++ lib/api.js
@@ -132,6 +132,11 @@ const api = {
     const json = await response.json()
     return json.data.patients
   },
+  async getDentists () {
+    const response = await fetch(`${BASE_URL}/dentists`)
+    const json = await response.json()
+    return json.data.dentists
+  },
   async getDentistById (id) {
     const response = await fetch(`${BASE_URL}/dentists/${id}`)
     const json = await response.json()
